Add contract args to hardhat env for fork networks
