+++ .github/workflows/release.yml
@@ -34,17 +34,13 @@ jobs:
           aws-secret-access-key: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
           aws-region: ${{ env.AWS_REGION }}
 
-      - name: Install dependencies
+      - name: Install all package dependencies
         run: npm ci
 
-      - name: Build helpers
-        run: npm run build --workspace @casimir/helpers
-
-      - name: Build landing app
-        run: npm run build --workspace @casimir/landing
-
-      - name: Build auth service
-        run: npm run build --workspace @casimir/auth
+      - name: Build all packages
+        run: npm exec --workspaces -- npm run build
+        env:
+          PUBLIC_STAGE: ${{ env.STAGE }}
 
       - name: Deploy cdk infrastructure
         run: npm run deploy
