+++ apps/web/src/composables/environment.ts
@@ -1,5 +1,5 @@
 export default function useEnvironment() {
-    const ethereumURL = import.meta.env.PUBLIC_ETHEREUM_RPC || 'http://127.0.0.1:8545/'
+    const ethereumURL = import.meta.env.PUBLIC_ETHEREUM_URL || 'http://127.0.0.1:8545/'
     const ledgerType = import.meta.env.PUBLIC_SPECULOS_PORT ? 'speculos' : 'usb'
     const speculosURL = import.meta.env.PUBLIC_SPECULOS_PORT ? `http://127.0.0.1:${import.meta.env.PUBLIC_SPECULOS_PORT}` : undefined
     const walletConnectURL = 'https://bridge.walletconnect.org'
