+++ common/data/package.json
@@ -4,14 +4,12 @@
     "main": "src/index.ts",
     "scripts": {
         "configure:python": "poetry install && poetry run ipython kernel install --user --name=casimir-data",
-        "dev": "npx esno src/index.ts",
         "build": "esbuild src/index.ts --bundle --minify --sourcemap --platform=node --target=es2020 --outfile=dist/index.js",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "devDependencies": {
         "@types/node": "^17.0.38",
-        "esbuild": "^0.15.9",
-        "esno": "^0.16.3"
+        "esbuild": "^0.15.9"
     },
     "dependencies": {
         "@aws-cdk/aws-glue-alpha": "^2.33.0-alpha.0"
