+++ common/ethers-helpers/package.json
@@ -3,14 +3,14 @@
     "private": true,
     "main": "src/index.ts",
     "scripts": {
-        "dev": "npx esno src/index.ts",
         "build": "esbuild src/index.ts --bundle --minify --sourcemap --platform=node --target=esnext --outfile=dist/index.js",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
+    "dependencies": {
+        "ethers": "^5.6.9"
+    },
     "devDependencies": {
         "@types/node": "^17.0.38",
-        "esbuild": "^0.15.9",
-        "esno": "^0.16.3",
-        "ethers": "^5.6.9"
+        "esbuild": "^0.15.9"
     }
 }
