+++ common/hardhat-helpers/package.json
@@ -3,8 +3,7 @@
     "private": true,
     "main": "src/index.ts",
     "scripts": {
-        "dev": "npx esno src/index.ts",
-        "build": "esbuild src/index.ts --bundle --minify --sourcemap --platform=node --target=esnext --outfile=dist/index.js",
+        "build": "echo 'Warning: @casimir/hardhat-helpers build not specified. Disregard this warning and any listed errors above if @casimir/hardhat-helpers is not needed for the current project build.' && exit 0",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "devDependencies": {
@@ -13,7 +12,6 @@
         "@typechain/hardhat": "^6.1.2",
         "@types/node": "^17.0.38",
         "esbuild": "^0.15.9",
-        "esno": "^0.16.3",
         "hardhat": "^2.12.2",
         "typechain": "^8.1.0"
     }
