+++ common/string-helpers/.gitignore
