+++ common/string-helpers/package.json
@@ -1,15 +1,13 @@
 {
-  "name": "@casimir/helpers",
+  "name": "@casimir/string-helpers",
   "private": true,
   "main": "src/index.ts",
   "scripts": {
-    "dev": "npx esno src/index.ts",
     "build": "esbuild src/index.ts --bundle --minify --sourcemap --platform=node --target=es2020 --outfile=dist/index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "devDependencies": {
     "@types/node": "^17.0.38",
-    "esbuild": "^0.15.9",
-    "esno": "^0.16.3"
+    "esbuild": "^0.15.9"
   }
 }
