+++ common/transport-speculos-http/package.json
@@ -2,12 +2,11 @@
   "name": "@casimir/transport-speculos-http",
   "private":"true",
   "main": "src/index.ts",
-  "dependencies": {
-    "@ledgerhq/hw-transport": "^6.27.2"
-  },
   "scripts": {
-    "dev": "npx esno src/index.ts",
     "build": "esbuild src/index.ts --bundle --minify --sourcemap --platform=node --target=esnext --outfile=dist/index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "dependencies": {
+    "@ledgerhq/hw-transport": "^6.27.2"
   }
 }
