+++ common/types/package.json
@@ -1,5 +1,8 @@
 {
     "name": "@casimir/types",
     "private": true,
-    "main": "src/index.ts"
+    "main": "src/index.ts",
+    "scripts": {
+        "build": "echo 'Warning: @casimir/types build not specified. Disregard this warning and any listed errors above if @casimir/types is not needed for the current project build.' && exit 0"
+    }
 }
