+++ contracts/ethereum/docs/index.md
@@ -2,6 +2,25 @@
 
 ## SSVManager
 
+### Token
+
+```solidity
+enum Token {
+  LINK,
+  SSV,
+  WETH
+}
+```
+
+### Fees
+
+```solidity
+struct Fees {
+  uint256 LINK;
+  uint256 SSV;
+}
+```
+
 ### UserAccount
 
 ```solidity
@@ -11,6 +30,14 @@ struct UserAccount {
 }
 ```
 
+### tokens
+
+```solidity
+mapping(enum SSVManager.Token => address) tokens
+```
+
+Token addresses
+
 ### users
 
 ```solidity
@@ -18,14 +45,6 @@ mapping(address => struct SSVManager.UserAccount) users
 ```
 
 All users who have deposited to pools
-
-### depositFee
-
-```solidity
-uint256 depositFee
-```
-
-Current deposit fee
 
 ### openPools
 
@@ -54,7 +73,7 @@ Event signaling a user deposit to a pool
 ### constructor
 
 ```solidity
-constructor() public
+constructor(address linkTokenAddress, address ssvTokenAddress, address wethTokenAddress) public
 ```
 
 ### deposit
@@ -65,13 +84,37 @@ function deposit() external payable
 
 Deposit to the pool manager
 
+### processFees
+
+```solidity
+function processFees(uint256 depositAmount) private returns (uint256, uint256, uint256)
+```
+
+_Process fees from deposit_
+
+#### Return Values
+
+| Name | Type | Description |
+| ---- | ---- | ----------- |
+| [0] | uint256 | The remaining deposit amount after fees |
+| [1] | uint256 |  |
+| [2] | uint256 |  |
+
+### swap
+
+```solidity
+function swap(address tokenIn, address tokenOut, uint256 amountIn) private returns (uint256)
+```
+
+_Swap one token for another_
+
 ### deployPool
 
 ```solidity
 function deployPool(bytes32 _salt) private returns (address)
 ```
 
-_Deploys a new pool contract_
+_Deploy a new pool contract_
 
 #### Return Values
 
@@ -79,13 +122,47 @@ _Deploys a new pool contract_
 | ---- | ---- | ----------- |
 | [0] | address | The address of the newly deployed pool contract |
 
-### getDepositFee
+### getFees
 
 ```solidity
-function getDepositFee() external view returns (uint256)
+function getFees() public pure returns (struct SSVManager.Fees)
 ```
 
-Get the current deposit fee
+Get the current token fees as percentages
+
+#### Return Values
+
+| Name | Type | Description |
+| ---- | ---- | ----------- |
+| [0] | struct SSVManager.Fees | The current token fees as percentages |
+
+### getLINKFee
+
+```solidity
+function getLINKFee() public pure returns (uint256)
+```
+
+Get the LINK fee percentage to charge on each deposit
+
+#### Return Values
+
+| Name | Type | Description |
+| ---- | ---- | ----------- |
+| [0] | uint256 | The LINK fee percentage to charge on each deposit |
+
+### getSSVFee
+
+```solidity
+function getSSVFee() public pure returns (uint256)
+```
+
+Get the SSV fee percentage to charge on each deposit
+
+#### Return Values
+
+| Name | Type | Description |
+| ---- | ---- | ----------- |
+| [0] | uint256 | The SSV fee percentage to charge on each deposit |
 
 ### getOpenPools
 
@@ -239,5 +316,103 @@ function getBalance() external view returns (uint256)
 
 ```solidity
 function getUserBalance(address userAddress) external view returns (uint256)
+```
+
+## WETH9
+
+### name
+
+```solidity
+string name
+```
+
+### symbol
+
+```solidity
+string symbol
+```
+
+### decimals
+
+```solidity
+uint8 decimals
+```
+
+### Approval
+
+```solidity
+event Approval(address src, address guy, uint256 wad)
+```
+
+### Transfer
+
+```solidity
+event Transfer(address src, address dst, uint256 wad)
+```
+
+### Deposit
+
+```solidity
+event Deposit(address dst, uint256 wad)
+```
+
+### Withdrawal
+
+```solidity
+event Withdrawal(address src, uint256 wad)
+```
+
+### balanceOf
+
+```solidity
+mapping(address => uint256) balanceOf
+```
+
+### allowance
+
+```solidity
+mapping(address => mapping(address => uint256)) allowance
+```
+
+### 
+
+```solidity
+undefined() external payable
+```
+
+### 
+
+```solidity
+undefined() public payable
+```
+
+### 
+
+```solidity
+undefined(uint256 wad) public
+```
+
+### 
+
+```solidity
+undefined() public view returns (uint256)
+```
+
+### 
+
+```solidity
+undefined(address guy, uint256 wad) public returns (bool)
+```
+
+### 
+
+```solidity
+undefined(address dst, uint256 wad) public returns (bool)
+```
+
+### 
+
+```solidity
+undefined(address src, address dst, uint256 wad) public returns (bool)
 ```
