+++ contracts/ethereum/scripts/ssv.deploy.ts
@@ -3,9 +3,9 @@ import { deployContract } from '@casimir/hardhat-helpers'
 void async function () {
     const name = 'SSVManager'
     const args = {
-        linkTokenAddress: process.env.LINK_TOKEN_ADDRESS,
-        ssvTokenAddress: process.env.SSV_TOKEN_ADDRESS,
-        wethTokenAddress: process.env.WETH_TOKEN_ADDRESS
+        LINKTokenAddress: process.env.LINK_TOKEN_ADDRESS,
+        SSVTokenAddress: process.env.SSV_TOKEN_ADDRESS,
+        WETHTokenAddress: process.env.WETH_TOKEN_ADDRESS
     }
     const options = {}
     const proxy = false
