--- contracts/ethereum/src/SSVToken.sol
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-or-later
-pragma solidity 0.8.4;
-
-import '../scripts/resources/ssv-contracts/contracts/token/SSVToken.sol';
