+++ contracts/ethereum/src/external/DepositContract.sol
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+pragma solidity 0.6.11;
+
+import '../../scripts/resources/consensus-specs/solidity_deposit_contract/deposit_contract.sol';
