+++ contracts/ethereum/src/external/SSVRegistry.sol
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+pragma solidity ^0.8.2;
+
+import '../../scripts/resources/ssv-network/contracts/SSVRegistry.sol';
