+++ contracts/ethereum/src/external/SSVToken.sol
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+pragma solidity 0.8.4;
+
+import '../../scripts/resources/ssv-contracts/contracts/token/SSVToken.sol';
