--- contracts/ethereum/src/interfaces/ISSVManager.sol
@@ -1,2 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-or-later
-pragma solidity 0.8.16;
