+++ scripts/local/dev
@@ -84,7 +84,7 @@ else
 
             CHAIN=$(echo $chain | tr '[:lower:]' '[:upper:]')
             rpc_base=${chain}_${network}
-            export "PUBLIC_${CHAIN}_RPC"="${!rpc_base}/$rpc_key"
+            export "PUBLIC_${CHAIN}_URL"="${!rpc_base}/$rpc_key"
         else
             # Pass fork or network to chain-specific script
             commands+=("npm run dev:$chain --fork=$fork")
